refactored tests for PublicAPIRouter
