+++ spec/ValidationAndPasswordsReset.spec.js
@@ -1,10 +1,8 @@
 'use strict';
 
-const { PublicAPIRouter, pages } = require('../lib/Routers/PublicAPIRouter');
 const MockEmailAdapterWithOptions = require('./MockEmailAdapterWithOptions');
 const request = require('../lib/request');
 const Config = require('../lib/Config');
-const Utils = require('../lib/Utils');
 
 describe('Custom Pages, Email Verification, Password Reset', () => {
   it('should set the custom pages', done => {
@@ -1076,150 +1074,5 @@ describe('Custom Pages, Email Verification, Password Reset', () => {
         fail(JSON.stringify(error));
         done();
       });
-  });
-
-  describe('localization of custom pages', () => {
-    let router = new PublicAPIRouter();
-    let req;
-    let pageResponse;
-    let redirectResponse;
-    const config = {
-      appId: 'test',
-      appName: 'ExampleAppName',
-      verifyUserEmails: true,
-      emailAdapter: {
-        sendVerificationEmail: () => Promise.resolve(),
-        sendPasswordResetEmail: () => Promise.resolve(),
-        sendMail: () => {},
-      },
-      publicServerURL: 'http://localhost:8378/1',
-      enablePageLocalization: true,
-    };
-
-    beforeEach(async () => {
-      router = new PublicAPIRouter();
-      pageResponse = spyOn(router, 'pageResponse').and.callThrough();
-      redirectResponse = spyOn(router, 'redirectResponse').and.callThrough();
-      req = {
-        method: 'GET',
-        config: {
-          customPages: {},
-          enablePageLocalization: true,
-          publicServerURL: 'http://example.com',
-        },
-        query: {
-          locale: 'de-AT',
-        },
-      };
-    });
-
-    it('returns default file if localization is disabled', async () => {
-      delete req.config.enablePageLocalization;
-
-      await expectAsync(router.goToPage(req, pages.invalidLink)).toBeResolved();
-      expect(pageResponse.calls.all()[0].args[1]).toBeDefined();
-      expect(pageResponse.calls.all()[0].args[1]).not.toMatch(
-        new RegExp(`\/de(-AT)?\/${pages.invalidLink.defaultFile}`)
-      );
-    });
-
-    it('returns default file if no locale is specified', async () => {
-      delete req.query.locale;
-
-      await expectAsync(router.goToPage(req, pages.invalidLink)).toBeResolved();
-      expect(pageResponse.calls.all()[0].args[1]).toBeDefined();
-      expect(pageResponse.calls.all()[0].args[1]).not.toMatch(
-        new RegExp(`\/de(-AT)?\/${pages.invalidLink.defaultFile}`)
-      );
-    });
-
-    it('returns custom page regardless of localization enabled', async () => {
-      req.config.customPages = { invalidLink: 'http://invalid-link.example.com' };
-
-      await expectAsync(router.goToPage(req, pages.invalidLink)).toBeResolved();
-      expect(pageResponse).not.toHaveBeenCalled();
-      expect(redirectResponse.calls.all()[0].args[0]).toBe(req.config.customPages.invalidLink);
-    });
-
-    it('returns file for locale match', async () => {
-      await expectAsync(router.goToPage(req, pages.invalidLink)).toBeResolved();
-      expect(pageResponse.calls.all()[0].args[1]).toBeDefined();
-      expect(pageResponse.calls.all()[0].args[1]).toMatch(
-        new RegExp(`\/de-AT\/${pages.invalidLink.defaultFile}`)
-      );
-    });
-
-    it('returns file for language match', async () => {
-      // Pretend no locale matching file exists
-      spyOn(Utils, 'fileExists').and.callFake(async path => {
-        return !path.includes(`/de-AT/${pages.invalidLink.defaultFile}`);
-      });
-
-      await expectAsync(router.goToPage(req, pages.invalidLink)).toBeResolved();
-      expect(pageResponse.calls.all()[0].args[1]).toBeDefined();
-      expect(pageResponse.calls.all()[0].args[1]).toMatch(
-        new RegExp(`\/de\/${pages.invalidLink.defaultFile}`)
-      );
-    });
-
-    it('returns default file for neither locale nor language match', async () => {
-      req.query.locale = 'yo-LO';
-
-      await expectAsync(router.goToPage(req, pages.invalidLink)).toBeResolved();
-      expect(pageResponse.calls.all()[0].args[1]).toBeDefined();
-      expect(pageResponse.calls.all()[0].args[1]).not.toMatch(
-        new RegExp(`\/yo(-LO)?\/${pages.invalidLink.defaultFile}`)
-      );
-    });
-
-    it('returns a file for GET request', async () => {
-      await expectAsync(router.goToPage(req, pages.invalidLink)).toBeResolved();
-      expect(pageResponse).toHaveBeenCalled();
-      expect(redirectResponse).not.toHaveBeenCalled();
-    });
-
-    it('returns a redirect for POST request', async () => {
-      req.method = 'POST';
-      await expectAsync(router.goToPage(req, pages.invalidLink)).toBeResolved();
-      expect(pageResponse).not.toHaveBeenCalled();
-      expect(redirectResponse).toHaveBeenCalled();
-    });
-
-    it('returns a redirect for custom pages for GET and POST', async () => {
-      req.config.customPages = { invalidLink: 'http://invalid-link.example.com' };
-
-      for (const method of ['GET', 'POST']) {
-        req.method = method;
-        await expectAsync(router.goToPage(req, pages.invalidLink)).toBeResolved();
-        expect(pageResponse).not.toHaveBeenCalled();
-        expect(redirectResponse).toHaveBeenCalled();
-      }
-    });
-
-    it('responds to POST request with redirect response (e2e test)', async () => {
-      await reconfigureServer(config);
-      const response = await request({
-        url:
-          'http://localhost:8378/1/apps/test/request_password_reset?token=exampleToken&username=exampleUsername&locale=de-AT',
-        followRedirects: false,
-        method: 'POST',
-      });
-      expect(response.status).toEqual(303);
-      expect(response.headers.location).toEqual(
-        'http://localhost:8378/apps/de-AT/invalid_link.html'
-      );
-    });
-
-    it('responds to GET request with content response (e2e test)', async () => {
-      await reconfigureServer(config);
-      const response = await request({
-        url:
-          'http://localhost:8378/1/apps/test/request_password_reset?token=exampleToken&username=exampleUsername&locale=de-AT',
-        followRedirects: false,
-        method: 'GET',
-      });
-      expect(response.status).toEqual(200);
-      expect(response.text).toContain('<html>');
-    });
   });
 });
